Add more logic for controlling when modules fade in and out
The modules will be faded in when the PIR sensor is triggered,
and a timer will start when the PIR sensor signals "no user
presence". The modules will be hidden when the timer expires.

This module can also control the monitor, by enabling/disabling
HDMI output on the Pi. The monitor is turned on by PIR input and
is turned off after a timer expires.

A node_helper.js file was created to handle running the commands
that turn the monitor on and off.

The duration of events is configurable.

I added many comments (probably too many) so I can figure
out what I was thinking in the future.